Slack apps: move to Build menu, list/detail/create/edit views
Reworks Slack apps from a single inline form-and-list panel under Settings into
a first-class Build-area feature modeled on /agents.

- Nav: "Slack apps" moves from Settings into the Build menu, right after
  Automations (IconSlack); removed from the settings sub-nav. Lives at
  /<ws>/slack-apps.
- List (/slack-apps): one clickable row per app (name, status, labels) like the
  agents inventory, with a "+ New Slack app" button top-right. Admin-only.
- Create (/slack-apps/new): the create form in its own view like /agents/new
  (back link, header); redirects to the new app's detail on success.
- Detail (/slack-apps/[id]): read view — status, owner, labels, setup steps,
  request URLs, app manifest (show/copy), and the Add-to-Slack/Reinstall link —
  with an "Edit" button top-right and the install-result banner.
- Edit (/slack-apps/[id]/edit): the credentials/config form + delete in its own
  view; redirects to detail on save, to the list on delete.

Wiring: the create/update actions now redirect to the detail view on success
(error stays inline); the OAuth install callback bounces back to the app's
detail page instead of Settings; old /settings/slack 302-redirects to
/slack-apps for stale bookmarks. The audit writes added earlier are preserved.

tsc + eslint clean; 101 tests pass.
