+++ web/src/app/[workspace]/settings/settings-nav.tsx
@@ -18,7 +18,6 @@ const ITEMS: Item[] = [
   { slug: "composio", label: "Composio" },
   { slug: "tembo", label: "Tembo Coding Agent" },
   { slug: "api-keys", label: "API keys" },
-  { slug: "slack", label: "Slack apps" },
   { slug: "appearance", label: "Appearance" },
   { slug: "deleted-agents", label: "Deleted agents" },
   { slug: "version", label: "Version" },
